fix: cleaned up a bit in the code
