+++ listing/index.html
@@ -79,6 +79,12 @@
                   Bid
                 </button>
               </form>
+              <div id="auctionClosed" class="text-2xl text-red-700 hidden">
+                This Auction is closed!
+              </div>
+              <div id="wrongBid" class="text-2xl text-red-700 hidden">
+                Your bid needs to be higher than the last bid.
+              </div>
             </div>
           </div>
           <div class="flex flex-col w-full max-w-[500px] gap-6 items-center">
