+++ profile/index.html
@@ -33,7 +33,7 @@
         >
           <div class="w-full max-w-[400px]">
             <h2 class="text-center text-2xl font-semibold my-4">
-              Your latest listings
+              Latest listings
             </h2>
             <div
               id="activeListings"
@@ -43,7 +43,7 @@
 
           <div class="w-full max-w-[400px]">
             <h2 class="text-center text-2xl font-semibold my-4">
-              Listings about to close
+              Closed listings
             </h2>
             <div
               id="closedListings"
