+++ src/ts/ui/listing/bid.ts
@@ -1,8 +1,14 @@
 import { makeBid } from '../../api/listing/bid';
+import { ListingObject } from '../../utilities/types';
 
-export const onBid = (event: SubmitEvent, id: string) => {
+export const onBid = (
+  event: SubmitEvent,
+  id: string,
+  listing: ListingObject
+) => {
   event.preventDefault();
   const form = event.target as HTMLFormElement;
+  const bidError = document.getElementById('wrongBid');
 
   const formData = new FormData(form);
   const numberInput = formData.get('bid');
@@ -13,7 +19,11 @@ export const onBid = (event: SubmitEvent, id: string) => {
     return;
   }
 
-  console.log(`bidding this amount: ${amount} on this id: `, id);
-
+  if (listing.bids[0]) {
+    if (listing.bids[0].amount >= amount) {
+      bidError?.classList.remove('hidden');
+      return;
+    }
+  }
   makeBid(id, amount);
 };
